Fix using minute for x-Axis
Recently, the option to (not) plot the time axis using minutes was
introduced.
This commit fixes some methods that were not properly implemented.
Moreover, the name of the flag was changed from `use_minutes` to
`x_axis_in_minutes` to make clear that only plotting is affected and not
other parameter values (e.g. start and end times).